modified Author (id: int->Long)

+++ src/main/java/ru/otus/spring/dao/AuthorDao.java
@@ -10,11 +10,11 @@ public interface AuthorDao {
 	
 	void insert(Author author);
 	
-	Author getById(int id);
+	Author getById(Long id);
 	
 	List<Author> getAll();
 	
-	void deleteById(int id);
+	void deleteById(Long id);
 
 	Author getByName(String name);
