+++ src/main/java/ru/otus/spring/mapper/BookMapper.java
@@ -17,7 +17,7 @@ public class BookMapper implements RowMapper<Book> {
 	public Book mapRow(ResultSet rs, int rowNum) throws SQLException {
 		final int id = rs.getInt("id");
 		final String name = rs.getString("name");
-		final int author_id = rs.getInt("author_id");
+		final Long author_id = rs.getLong("author_id");
 		final String author = rs.getString("author");
 		final int genre_id = rs.getInt("genre_id");
 		final String genre = rs.getString("genre");
